proyecto02/requirements.txt
PySide6>=6.5.0
# Librerías principales
PySide6==6.8.1.1
python-docx==1.1.2
docx2pdf==0.1.8
lxml==5.2.1
